TPCClusterFinder: Add syntactic sugar for getting the position of a digit.
