+++ GPU/GPUTracking/TPCClusterFinder/GPUTPCSharedMemoryData.h
@@ -15,6 +15,7 @@
 #define O2_GPU_GPU_TPC_SHARED_MEMORY_DATA_H
 
 #include "clusterFinderDefs.h"
+#include "ChargePos.h"
 #include "PackedCharge.h"
 
 #ifndef SCRATCH_PAD_WORK_GROUP_SIZE
